Refactor utp-testing crate
