+++ .circleci/config.yml
@@ -124,7 +124,7 @@ jobs:
           command: docker build -t test-utp --no-cache utp-testing/docker/circleci/
       - run:
           name: Run Test App with network simulator
-          command: SCENARIO="drop-rate --delay=15ms --bandwidth=10Mbps --queue=25 --rate_to_client=10 --rate_to_server=10" docker-compose -f utp-testing/docker/docker-compose.yml up -d
+          command: SCENARIO="drop-rate --delay=15ms --bandwidth=10Mbps --queue=25 --rate_to_client=0 --rate_to_server=0" docker-compose -f utp-testing/docker/docker-compose.yml up -d
       - run:
           name: Wait all containers to start
           command: sleep 5
