+++ utp-testing/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0.52"
 jsonrpsee = {version = "0.15.1", features = ["full"]}
+rand = "0.8.5"
 structopt = "0.3.26"
 tracing = "0.1.26"
 tracing-subscriber = "0.2.18"
@@ -14,10 +15,6 @@ tokio = {version = "1.8.0", features = ["full"]}
 discv5 = { git = "https://github.com/sigp/discv5.git", branch = "master" }
 hex = "0.4.3"
 log = "0.4.14"
-
-[[bin]]
-name = "utp-testing"
-path = "src/bin/local_test.rs"
 
 [[bin]]
 name = "utp-test-app"
